+++ src/renderer/components/basemapList/BasemapListItem.js
